requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==3.2.23
django-allauth==0.58.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
sqlparse==0.4.4
whitenoise==6.6.0
